Switched from laravel-modules plugin to native Laravel Packages
