--- Modules/MediasourceFile/Config/config.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'name' => 'MediasourceFile'
-];
