--- Modules/MediasourceFile/Models/MediaSource.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Modules\MediasourceFile\Models;
-
-class MediaSource extends \App\MediaSource {
-
-    const identifier = 'file';
-
-    const icon = ['fas', 'music'];
-
-}
