--- Modules/MediasourceFile/composer.json
@@ -1,25 +1,0 @@
-{
-    "name": "nwidart/mediasourcefile",
-    "description": "",
-    "authors": [
-        {
-            "name": "Nicolas Widart",
-            "email": "n.widart@gmail.com"
-        }
-    ],
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Modules\\MediasourceFile\\Providers\\MediasourceFileServiceProvider"
-            ],
-            "aliases": {
-
-            }
-        }
-    },
-    "autoload": {
-        "psr-4": {
-            "Modules\\MediasourceFile\\": ""
-        }
-    }
-}
