--- Modules/MediasourceFile/module.json
@@ -1,13 +1,0 @@
-{
-    "name": "MediasourceFile",
-    "alias": "mediasourcefile",
-    "description": "",
-    "keywords": [],
-    "order": 0,
-    "providers": [
-        "Modules\\MediasourceFile\\Providers\\MediasourceFileServiceProvider"
-    ],
-    "aliases": {},
-    "files": [],
-    "requires": []
-}
