--- Modules/ShoutzorAcoustId/Providers/ShoutzorAcoustIdServiceProvider.php
@@ -1,80 +1,0 @@
-<?php
-
-namespace Modules\ShoutzorAcoustId\Providers;
-
-use Illuminate\Support\ServiceProvider;
-
-class ShoutzorAcoustIdServiceProvider extends ServiceProvider
-{
-    /**
-     * @var string $moduleName
-     */
-    protected $moduleName = 'ShoutzorAcoustId';
-
-    /**
-     * @var string $moduleNameLower
-     */
-    protected $moduleNameLower = 'shoutzoracoustid';
-
-    /**
-     * Boot the application events.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        $this->registerTranslations();
-        $this->registerConfig();
-    }
-
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
-    {
-
-    }
-
-    /**
-     * Register config.
-     *
-     * @return void
-     */
-    protected function registerConfig()
-    {
-        $this->publishes([
-            module_path($this->moduleName, 'Config/config.php') => config_path($this->moduleNameLower . '.php'),
-        ], 'config');
-        $this->mergeConfigFrom(
-            module_path($this->moduleName, 'Config/config.php'), $this->moduleNameLower
-        );
-    }
-
-    /**
-     * Register translations.
-     *
-     * @return void
-     */
-    public function registerTranslations()
-    {
-        $langPath = resource_path('lang/modules/' . $this->moduleNameLower);
-
-        if (is_dir($langPath)) {
-            $this->loadTranslationsFrom($langPath, $this->moduleNameLower);
-        } else {
-            $this->loadTranslationsFrom(module_path($this->moduleName, 'Resources/lang'), $this->moduleNameLower);
-        }
-    }
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return [];
-    }
-}
