--- Modules/ShoutzorAcoustId/composer.json
@@ -1,25 +1,0 @@
-{
-    "name": "nwidart/shoutzoracoustid",
-    "description": "",
-    "authors": [
-        {
-            "name": "Nicolas Widart",
-            "email": "n.widart@gmail.com"
-        }
-    ],
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Modules\\ShoutzorAcoustId\\Providers\\ShoutzorAcoustIdServiceProvider"
-            ],
-            "aliases": {
-
-            }
-        }
-    },
-    "autoload": {
-        "psr-4": {
-            "Modules\\ShoutzorAcoustId\\": ""
-        }
-    }
-}
