--- Modules/ShoutzorAcoustId/module.json
@@ -1,14 +1,0 @@
-{
-    "name": "ShoutzorAcoustId",
-    "alias": "shoutzoracoustid",
-    "description": "Uses AcoustId for identifying audio files",
-    "keywords": [],
-    "order": 0,
-    "providers": [
-        "Modules\\ShoutzorAcoustId\\Providers\\ShoutzorAcoustIdServiceProvider",
-        "Modules\\ShoutzorAcoustId\\Providers\\EventServiceProvider"
-    ],
-    "aliases": {},
-    "files": [],
-    "requires": []
-}
