+++ composer.json
@@ -6,7 +6,7 @@
         "shoutz0r",
         "shoutzor"
     ],
-    "license": "GPLv3",
+    "license": "GPL-3.0-only",
     "require": {
         "php": "^7.4",
         "ext-json": "*",
@@ -17,18 +17,18 @@
         "laravel/passport": "^9.2",
         "laravel/tinker": "^2.0",
         "laravel/ui": "^2.0",
-        "nwidart/laravel-modules": "^7.1",
         "symfony/event-dispatcher": "^5.1",
         "ext-curl": "*"
     },
     "require-dev": {
         "facade/ignition": "^2.0",
         "fzaninotto/faker": "^1.9.1",
-        "rauwebieten/php-faker-music": "^1.0",
         "mmo/faker-images": "^0.1.0",
         "mockery/mockery": "^1.3.1",
         "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "orchestra/testbench": "^5.0",
+        "phpunit/phpunit": "^8.5",
+        "rauwebieten/php-faker-music": "^1.0"
     },
     "config": {
         "optimize-autoloader": true,
