+++ packages/Shoutz0r/AcoustId/composer.json
@@ -1,0 +1,32 @@
+{
+    "name": "Shouzt0r/AcoustId",
+    "description": "Enables AcoustId for identifying audio files on Shoutz0r",
+    "type": "project",
+    "license": "GPL-3.0-only",
+    "authors": [{
+        "name": "Lukáš Lalinský",
+        "homepage": "https://acoustid.org/",
+        "role": "Developer"
+    },{
+        "name": "Jorin Vermeulen",
+        "homepage": "https://github.com/xorinzor/Shoutz0r-webui/",
+        "role": "Developer"
+    }],
+    "autoload": {
+        "psr-4": {
+            "Shoutz0r\\AcoustId\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Shoutz0r\\AcoustId\\Tests\\": "tests/"
+        }
+    },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "Shoutz0r\\AcoustId\\Providers\\PackageServiceProvider"
+            ]
+        }
+    }
+}
