+++ packages/Shoutz0r/AcoustId/config/config.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'shoutzor.acoustid.apikey'  => env('SHOUTZOR_ACOUSTID_APIKEY', ''),
+];
