+++ packages/Shoutz0r/LastFM/config/config.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'shoutzor.lastfm.apikey'  => env('SHOUTZOR_LASTFM_APPKEY', ''),
+    'shoutzor.lastfm.secret'  => env('SHOUTZOR_LASTFM_SECRET', ''),
+];
