+++ packages/Shoutz0r/LastFM/src/Subscribers/UploadSubscriber.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Shoutz0r\LastFM\Listeners;
+
+use App\Events\Internal\UploadProcessingEvent;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+
+class UploadSubscriber implements EventSubscriberInterface
+{
+
+    public static function getSubscribedEvents()
+    {
+        return [
+            UploadProcessingEvent::NAME => [
+                ['onProcessUpload', 0]
+            ]
+        ];
+    }
+
+    /**
+     * Handle the event.
+     *
+     * @param  UploadProcessingEvent  $event
+     * @return void
+     */
+    public function onProcessUpload(UploadProcessingEvent $event)
+    {
+    }
+}
